refactor: merge classes containing check logic and classes scheduling them
The convention for repeated checks is to have a single class:
- implementing the check logic
- scheduling its repeated execution

In some cases, two classes were used to do this, so a refactoring is
necessary to follow the convention.

In this commit, the logic from `*CheckerScheduler` classes (i.e. the
ones scheduling the repeated executions) is just moved to the
`*CheckScheduler` classes (previously `*Checker`, i.e. the ones
implementing the check logic).
The `*CheckerScheduler` classes are removed and their usages replaced by
the `*CheckScheduler` classes.

(cherry picked from commit d1298802c059b5e5e2d94e8e90796641d3bf218b)
