+++ zeebe/engine/src/main/java/io/camunda/zeebe/engine/processing/message/PendingMessageSubscriptionCheckScheduler.java
@@ -9,10 +9,20 @@ package io.camunda.zeebe.engine.processing.message;
 
 import io.camunda.zeebe.engine.processing.message.command.SubscriptionCommandSender;
 import io.camunda.zeebe.engine.state.immutable.PendingMessageSubscriptionState;
+import io.camunda.zeebe.engine.state.immutable.ScheduledTaskState;
 import io.camunda.zeebe.engine.state.message.MessageSubscription;
+import io.camunda.zeebe.stream.api.ReadonlyStreamProcessorContext;
+import io.camunda.zeebe.stream.api.StreamProcessorLifecycleAware;
+import java.time.Duration;
 import java.time.InstantSource;
+import java.util.function.Supplier;
 
-public final class PendingMessageSubscriptionCheckScheduler implements Runnable {
+public final class PendingMessageSubscriptionCheckScheduler
+    implements Runnable, StreamProcessorLifecycleAware {
+
+  public static final Duration SUBSCRIPTION_TIMEOUT = Duration.ofSeconds(10);
+  public static final Duration SUBSCRIPTION_CHECK_INTERVAL = Duration.ofSeconds(30);
+
   private final SubscriptionCommandSender commandSender;
   private final PendingMessageSubscriptionState state;
 
@@ -34,6 +44,17 @@ public final class PendingMessageSubscriptionCheckScheduler implements Runnable 
     this.state = state;
     this.subscriptionTimeout = subscriptionTimeout;
     this.clock = clock;
+  }
+
+  public PendingMessageSubscriptionCheckScheduler(
+      final Supplier<ScheduledTaskState> scheduledTaskStateFactory,
+      final PendingMessageSubscriptionState pendingState,
+      final SubscriptionCommandSender subscriptionCommandSender,
+      final Duration messagesTtlCheckerInterval,
+      final int messagesTtlCheckerBatchLimit,
+      final boolean enableMessageTtlCheckerAsync,
+      final InstantSource clock) {
+    this(subscriptionCommandSender, pendingState, SUBSCRIPTION_TIMEOUT.toMillis(), clock);
   }
 
   @Override
@@ -59,5 +80,44 @@ public final class PendingMessageSubscriptionCheckScheduler implements Runnable 
     state.onSent(record, sentTime);
 
     return true; // to continue visiting
+  }
+
+  @Override
+  public void onRecovered(final ReadonlyStreamProcessorContext context) {
+    scheduleMessageTtlChecker(context);
+    schedulePendingMessageSubscriptionChecker(context);
+  }
+
+  private void scheduleMessageTtlChecker(final ReadonlyStreamProcessorContext context) {
+    /* NOT APPLICABLE HERE
+    final var scheduleService = context.getScheduleService();
+    final var messageState = scheduledTaskStateFactory.get().getMessageState();
+    final var timestamp = clock.millis() + messagesTtlCheckerInterval.toMillis();
+    final var timeToLiveChecker =
+        new MessageTimeToLiveCheckScheduler(
+            messagesTtlCheckerInterval,
+            messagesTtlCheckerBatchLimit,
+            enableMessageTtlCheckerAsync,
+            scheduleService,
+            messageState,
+            context.getClock());
+    if (enableMessageTtlCheckerAsync) {
+      scheduleService.runAtAsync(timestamp, timeToLiveChecker);
+    } else {
+      scheduleService.runAt(timestamp, timeToLiveChecker);
+    }
+    */
+  }
+
+  private void schedulePendingMessageSubscriptionChecker(
+      final ReadonlyStreamProcessorContext context) {
+    final var scheduleService = context.getScheduleService();
+    final var pendingSubscriptionChecker =
+        new PendingMessageSubscriptionCheckScheduler(
+            commandSender,
+            state,
+            SUBSCRIPTION_TIMEOUT.toMillis(),
+            context.getClock());
+    scheduleService.runAtFixedRate(SUBSCRIPTION_CHECK_INTERVAL, pendingSubscriptionChecker);
   }
 }
