+++ zeebe/engine/src/test/java/io/camunda/zeebe/engine/processing/message/MessageStreamProcessorTest.java
@@ -165,8 +165,8 @@ public final class MessageStreamProcessorTest {
             () -> {
               rule.getClock()
                   .addTime(
-                      MessageObserver.SUBSCRIPTION_CHECK_INTERVAL.plus(
-                          MessageObserver.SUBSCRIPTION_TIMEOUT));
+                      PendingMessageSubscriptionCheckScheduler.SUBSCRIPTION_CHECK_INTERVAL.plus(
+                          PendingMessageSubscriptionCheckScheduler.SUBSCRIPTION_TIMEOUT));
               verify(mockInterpartitionCommandSender, timeout(100).atLeast(2))
                   .sendCommand(
                       eq(0),
@@ -202,8 +202,8 @@ public final class MessageStreamProcessorTest {
             () -> {
               rule.getClock()
                   .addTime(
-                      MessageObserver.SUBSCRIPTION_CHECK_INTERVAL.plus(
-                          MessageObserver.SUBSCRIPTION_TIMEOUT));
+                      PendingMessageSubscriptionCheckScheduler.SUBSCRIPTION_CHECK_INTERVAL.plus(
+                          PendingMessageSubscriptionCheckScheduler.SUBSCRIPTION_TIMEOUT));
               verify(mockInterpartitionCommandSender, timeout(100).times(2))
                   .sendCommand(
                       eq(0),
